Progressivly worse ideas on dectree evaluation
A few different decision tree implementations. This code is incomplete,
but was enough for me to run some evaluations.

Locally, when compiled with avx enabled, the JNI version is about 20%
faster than naive on small trees, but only 8-10% faster as the trees
get larger (say, depth 8 w/ 500 trees). Batching the evaluation, to
take advantage of data parallelism and out of order evaluation, is
able to improve latency by 80% with a batch size of 16 on my laptop.
Batching is not yet implemented for the rescorer though, so this is
only theoretical.

* Convert trees into java bytecode with two different libraries. Both
  runtime compilers seem to fail on larger trees though. Faster than
  naive on small trees. Not really scalable and should probably be
  removed

* A branchless implementation that requires trees to be fully balanced
  in java. Stores data in ByteBuffer to keep the feature ids and
  thresholds close together in memory. Not faster than naive.

* Same branchless implementation, but using sun.misc.Unsafe. Probably
  a pretty bad idea. Does come up slightly faster than naive at times.

* Same branchless implementation, this time in C accessed via JNI. This
  seems to be faster than naive in most cases. It further improves
  when batching document evaluations.